requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.4
Flask-CORS==4.0.0
openai==0.28.0
whisper==1.1.10
SpeechRecognition==3.10.0
pydub==0.25.1
gtts==2.3.2
googletrans==4.0.0-rc1
PyPDF2==3.0.1
python-docx==0.8.11
Pillow==10.0.0
numpy==1.24.3
transformers==4.33.3
sentencepiece==0.1.99
pydantic==2.3.0
python-multipart==0.0.6
eventlet==0.33.3
python-dotenv==1.0.0
requests==2.31.0
moviepy==1.0.3
